Updated modularization
