+++ src/modules/Index/Resources.js
@@ -1,0 +1,13 @@
+import * as React from "react"
+import * as cls from "./classes"
+import Links from "./Links"
+import ReferenceManagers from "./ReferenceManagers"
+
+const Resources = () => (
+  <div>
+    <Links />
+    <ReferenceManagers />
+  </div>
+)
+
+export default Resources
