+++ src/pages/404.js
@@ -1,0 +1,1 @@
+export { default } from "../modules/NotFound/NotFoundPage";
