+++ src/styles/classes.js
@@ -1,5 +1,5 @@
 // Layout
-export const container = "mw7 center ph4"
+export const container = "mw9 center ph4"
 export const section = "pa4 bg-washed-blue br3 shadow-3 mb5"
 export const subsection = "pa4 bg-light-gray br2 shadow-1 mt4 mb4"
